Fix regex issue
The issue here is that the whitelist check was not being applied correctly. The negative lookahead (?!...) is checking if the next characters are not any of the whitelist words, but it's not considering word boundaries. This means that if "butt" is in the blacklist and "buttocks" is in the whitelist, the "butt" part of "buttocks" is still getting matched.